DM Audit: Final changes for database.yml recipe
Turns out that I don't need to check in the code each time I want
to push changes! This is the final update that works entirely
as expected. It will generate the identical 'production'
entry, just like before, and will now add in two additional MOI
entries. This can be modified if we ever need to hit anything else internal.